Add tests for setting utime
