+++ .gitignore
@@ -1,3 +1,7 @@
 /env/
 /__pycache__/
 mount/
+/.dir-locals.el
+/.run_tests.sh
+.cov2emacs.log
+/.coverage
